ERMAHGERD IT WORKS!!
The typechecker now gives the correct type when one defines recursive
higher order functions like foldr and map, no type signature needed!
This means fewer "It should do this, but doesn't" cop-outs in TC
explanation. Only known bug left is the fact that functions are still
polymorphic in their own definitions, but idk what to do about that or
what consequences it has.